backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
tensorflow==2.20.0
scikit-learn>=1.4.0
pandas==2.1.4
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
